# internal/swagger/v3/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : { },
    "description" : "A lightweight, flexible, elegant and full-featured RBAC scaffolding based on GIN + GORM 2.0 + Casbin 2.0 + Wire DI.",
    "title" : "gin-admin",
    "version" : "v10.1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/v1/menus" : {
      "get" : {
        "parameters" : [ {
          "description" : "Code path of menu (like xxx.xxx.xxx)",
          "in" : "query",
          "name" : "code",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Name of menu",
          "in" : "query",
          "name" : "name",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Whether to include menu resources",
          "in" : "query",
          "name" : "includeResources",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_api_v1_current_menus_get_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Query menu tree data",
        "tags" : [ "MenuAPI" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/schema.MenuForm"
              }
            }
          },
          "description" : "Request body",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_api_v1_menus_post_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Create menu record",
        "tags" : [ "MenuAPI" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/api/v1/menus/{id}" : {
      "delete" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Delete menu record by ID",
        "tags" : [ "MenuAPI" ]
      },
      "get" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_api_v1_menus_post_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Get menu record by ID",
        "tags" : [ "MenuAPI" ]
      },
      "put" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/schema.MenuForm"
              }
            }
          },
          "description" : "Request body",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Update menu record by ID",
        "tags" : [ "MenuAPI" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/api/v1/roles" : {
      "get" : {
        "parameters" : [ {
          "description" : "pagination index",
          "in" : "query",
          "name" : "current",
          "required" : true,
          "schema" : {
            "default" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "pagination size",
          "in" : "query",
          "name" : "pageSize",
          "required" : true,
          "schema" : {
            "default" : 10,
            "type" : "integer"
          }
        }, {
          "description" : "Display name of role",
          "in" : "query",
          "name" : "name",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Status of role (disabled, enabled)",
          "in" : "query",
          "name" : "status",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_api_v1_roles_get_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Query role list",
        "tags" : [ "RoleAPI" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/schema.RoleForm"
              }
            }
          },
          "description" : "Request body",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_api_v1_roles_post_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Create role record",
        "tags" : [ "RoleAPI" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/api/v1/roles/{id}" : {
      "delete" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Delete role record by ID",
        "tags" : [ "RoleAPI" ]
      },
      "get" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_api_v1_roles_post_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Get role record by ID",
        "tags" : [ "RoleAPI" ]
      },
      "put" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/schema.RoleForm"
              }
            }
          },
          "description" : "Request body",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Update role record by ID",
        "tags" : [ "RoleAPI" ],
        "x-codegen-request-body-name" : "body"
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "errors.Error" : {
        "properties" : {
          "code" : {
            "type" : "integer"
          },
          "detail" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "schema.Menu" : {
        "properties" : {
          "children" : {
            "description" : "Child menus",
            "items" : {
              "$ref" : "#/components/schemas/schema.Menu"
            },
            "type" : "array"
          },
          "code" : {
            "description" : "Code of menu (unique for each level)",
            "type" : "string"
          },
          "created_at" : {
            "description" : "Create time",
            "type" : "string"
          },
          "description" : {
            "description" : "Details about menu",
            "type" : "string"
          },
          "id" : {
            "description" : "Unique ID",
            "type" : "string"
          },
          "name" : {
            "description" : "Display name of menu",
            "type" : "string"
          },
          "parent_id" : {
            "description" : "Parent ID (From Menu.ID)",
            "type" : "string"
          },
          "parent_path" : {
            "description" : "Parent path (split by .)",
            "type" : "string"
          },
          "path" : {
            "description" : "Access path of menu",
            "type" : "string"
          },
          "properties" : {
            "description" : "Properties of menu (JSON)",
            "type" : "string"
          },
          "resources" : {
            "description" : "Resources of menu",
            "items" : {
              "$ref" : "#/components/schemas/schema.MenuResource"
            },
            "type" : "array"
          },
          "sequence" : {
            "description" : "Sequence for sorting (Order by desc)",
            "type" : "integer"
          },
          "status" : {
            "description" : "Status of menu (enabled, disabled)",
            "type" : "string"
          },
          "type" : {
            "description" : "Type of menu (page, button)",
            "type" : "string"
          },
          "updated_at" : {
            "description" : "Update time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "schema.MenuForm" : {
        "properties" : {
          "code" : {
            "description" : "Code of menu (unique for each level)",
            "maxLength" : 32,
            "type" : "string"
          },
          "description" : {
            "description" : "Details about menu",
            "type" : "string"
          },
          "name" : {
            "description" : "Display name of menu",
            "maxLength" : 128,
            "type" : "string"
          },
          "parent_id" : {
            "description" : "Parent ID (From Menu.ID)",
            "type" : "string"
          },
          "path" : {
            "description" : "Access path of menu",
            "type" : "string"
          },
          "properties" : {
            "description" : "Properties of menu (JSON)",
            "type" : "string"
          },
          "resources" : {
            "description" : "Resources of menu",
            "items" : {
              "$ref" : "#/components/schemas/schema.MenuResource"
            },
            "type" : "array"
          },
          "sequence" : {
            "description" : "Sequence for sorting (Order by desc)",
            "type" : "integer"
          },
          "status" : {
            "description" : "Status of menu (enabled, disabled)",
            "enum" : [ "disabled", "enabled" ],
            "type" : "string"
          },
          "type" : {
            "description" : "Type of menu (page, button)",
            "enum" : [ "page", "button" ],
            "type" : "string"
          }
        },
        "required" : [ "code", "name", "status", "type" ],
        "type" : "object"
      },
      "schema.MenuResource" : {
        "properties" : {
          "created_at" : {
            "description" : "Create time",
            "type" : "string"
          },
          "id" : {
            "description" : "Unique ID",
            "type" : "string"
          },
          "menu_id" : {
            "description" : "From Menu.ID",
            "type" : "string"
          },
          "method" : {
            "description" : "HTTP method",
            "type" : "string"
          },
          "path" : {
            "description" : "API request path (e.g. /api/v1/users/:id)",
            "type" : "string"
          },
          "updated_at" : {
            "description" : "Update time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "schema.Role" : {
        "properties" : {
          "code" : {
            "description" : "Code of role (unique)",
            "type" : "string"
          },
          "created_at" : {
            "description" : "Create time",
            "type" : "string"
          },
          "description" : {
            "description" : "Details about role",
            "type" : "string"
          },
          "id" : {
            "description" : "Unique ID",
            "type" : "string"
          },
          "menus" : {
            "description" : "Role menu list",
            "items" : {
              "$ref" : "#/components/schemas/schema.RoleMenu"
            },
            "type" : "array"
          },
          "name" : {
            "description" : "Display name of role",
            "type" : "string"
          },
          "sequence" : {
            "description" : "Sequence for sorting",
            "type" : "integer"
          },
          "status" : {
            "description" : "Status of role (disabled, enabled)",
            "type" : "string"
          },
          "updated_at" : {
            "description" : "Update time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "schema.RoleForm" : {
        "properties" : {
          "code" : {
            "description" : "Code of role (unique)",
            "maxLength" : 32,
            "type" : "string"
          },
          "description" : {
            "description" : "Details about role",
            "type" : "string"
          },
          "menus" : {
            "description" : "Role menu list",
            "items" : {
              "$ref" : "#/components/schemas/schema.RoleMenu"
            },
            "type" : "array"
          },
          "name" : {
            "description" : "Display name of role",
            "maxLength" : 128,
            "type" : "string"
          },
          "sequence" : {
            "description" : "Sequence for sorting",
            "type" : "integer"
          },
          "status" : {
            "description" : "Status of role (enabled, disabled)",
            "enum" : [ "disabled", "enabled" ],
            "type" : "string"
          }
        },
        "required" : [ "code", "name", "status" ],
        "type" : "object"
      },
      "schema.RoleMenu" : {
        "properties" : {
          "created_at" : {
            "description" : "Create time",
            "type" : "string"
          },
          "id" : {
            "description" : "Unique ID",
            "type" : "string"
          },
          "menu_id" : {
            "description" : "From Menu.ID",
            "type" : "string"
          },
          "role_id" : {
            "description" : "From Role.ID",
            "type" : "string"
          },
          "updated_at" : {
            "description" : "Update time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "util.ResponseResult" : {
        "properties" : {
          "data" : {
            "type" : "object"
          },
          "error" : {
            "$ref" : "#/components/schemas/errors.Error"
          },
          "success" : {
            "type" : "boolean"
          },
          "total" : {
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "_api_v1_current_menus_get_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/util.ResponseResult"
        }, {
          "properties" : {
            "data" : {
              "items" : {
                "$ref" : "#/components/schemas/schema.Menu"
              },
              "type" : "array"
            }
          },
          "type" : "object"
        } ]
      },
...
      "_api_v1_menus_post_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/util.ResponseResult"
        }, {
          "properties" : {
            "data" : {
              "$ref" : "#/components/schemas/schema.Menu"
            }
          },
          "type" : "object"
        } ]
      },
      "_api_v1_roles_get_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/util.ResponseResult"
        }, {
          "properties" : {
            "data" : {
              "items" : {
                "$ref" : "#/components/schemas/schema.Role"
              },
              "type" : "array"
            }
          },
          "type" : "object"
        } ]
      },
      "_api_v1_roles_post_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/util.ResponseResult"
        }, {
          "properties" : {
            "data" : {
              "$ref" : "#/components/schemas/schema.Role"
            }
          },
          "type" : "object"
        } ]
      },
...
    },
    "securitySchemes" : {
      "ApiKeyAuth" : {
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  },
...
}